# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentPassVault API
  description: Secure Secret Manager for Automated Agents
  version: 0.0.1
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/secrets:
    post:
      tags:
      - secret-controller
      operationId: createSecret
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SecretMetadataResponse"
  /api/v1/secrets/{secretId}/leases:
    get:
      tags:
      - lease-controller
      operationId: listLeases
      parameters:
      - name: secretId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: agentId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaseResponse"
    post:
      tags:
      - lease-controller
      operationId: createLease
      parameters:
      - name: secretId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/secrets/search:
    post:
      tags:
      - secret-controller
      operationId: searchSecrets
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchSecretRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretMetadataResponse"
  /api/v1/requests:
    post:
      tags:
      - request-controller
      operationId: createRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
  /api/v1/auth/reset-password:
    post:
      tags:
      - auth-controller
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/v1/auth/login/user:
    post:
      tags:
      - auth-controller
      operationId: userLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/v1/auth/login/agent:
    post:
      tags:
      - auth-controller
      operationId: agentLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/v1/auth/forgot-password:
    post:
      tags:
      - auth-controller
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/auth/change-password:
    post:
      tags:
      - auth-controller
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/agents:
    get:
      tags:
      - agent-controller
      operationId: listAgents
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentResponse"
    post:
      tags:
      - agent-controller
      operationId: createAgent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentTokenResponse"
  /api/v1/agents/{id}/rotate:
    post:
      tags:
      - agent-controller
      operationId: rotateToken
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentTokenResponse"
  /api/v1/agents/{id}/register:
    post:
      tags:
      - agent-controller
      operationId: registerAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterAgentRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/secrets/{id}:
    get:
      tags:
      - secret-controller
      operationId: getSecret
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: leaseToken
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SecretResponse"
    delete:
      tags:
      - secret-controller
      operationId: deleteSecret
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - secret-controller
      operationId: updateSecret
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecretRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SecretMetadataResponse"
  /api/v1/requests/{id}:
    get:
      tags:
      - request-controller
      operationId: getRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
    delete:
      tags:
      - request-controller
      operationId: abandonRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - request-controller
      operationId: updateRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
  /api/v1/auth/ping:
    get:
      tags:
      - auth-controller
      operationId: ping
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/agents/{id}:
    get:
      tags:
      - agent-controller
      operationId: getAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgentResponse"
    delete:
      tags:
      - agent-controller
      operationId: deleteAgent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/secrets/{secretId}/leases/{agentId}:
    delete:
      tags:
      - lease-controller
      operationId: revokeLease
      parameters:
      - name: secretId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: agentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    CreateSecretRequest:
      type: object
      properties:
        name:
          type: string
        encryptedValue:
          type: string
          maxLength: 87381
          minLength: 0
        metadata:
          type: object
          additionalProperties:
            type: object
      required:
      - encryptedValue
      - name
    SecretMetadataResponse:
      type: object
      properties:
        secretId:
          type: string
        name:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateLeaseRequest:
      type: object
      properties:
        agentId:
          type: string
          pattern: "^[0-9]+$"
        publicKey:
          type: string
        encryptedData:
          type: string
        expiry:
          type: string
          format: date-time
      required:
      - agentId
      - encryptedData
      - publicKey
    SearchSecretRequest:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties:
            type: object
    CreateRequestDTO:
      type: object
      properties:
        name:
          type: string
        context:
          type: string
          maxLength: 2048
          minLength: 0
        requiredMetadata:
          type: object
          additionalProperties:
            type: object
        requiredFieldsInSecretValue:
          type: array
          items:
            type: string
        type:
          type: string
          enum:
          - CREATE
          - LEASE
        secretId:
          type: string
          pattern: "^[0-9]+$"
      required:
      - name
      - type
    RequestResponse:
      type: object
      properties:
        requestId:
          type: string
        status:
          type: string
          enum:
          - pending
          - fulfilled
          - rejected
          - abandoned
        type:
          type: string
          enum:
          - CREATE
          - LEASE
        name:
          type: string
        context:
          type: string
        requiredMetadata:
          type: object
          additionalProperties:
            type: object
        requiredFieldsInSecretValue:
          type: array
          items:
            type: string
        mappedSecretId:
          type: string
        rejectionReason:
          type: string
        fulfillmentUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required:
      - newPassword
      - token
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
      - refreshToken
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
        refreshTokenExpiresIn:
          type: integer
          format: int64
    UserLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - password
      - username
    AgentLoginRequest:
      type: object
      properties:
        tenantId:
          type: string
          pattern: "^[0-9]+$"
        appToken:
          type: string
      required:
      - appToken
      - tenantId
    ForgotPasswordRequest:
      type: object
      properties:
        username:
          type: string
      required:
      - username
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
      - newPassword
      - oldPassword
    CreateAgentRequest:
      type: object
      properties:
        name:
          type: string
      required:
      - name
    AgentTokenResponse:
      type: object
      properties:
        agentId:
          type: string
        appToken:
          type: string
    RegisterAgentRequest:
      type: object
      properties:
        publicKey:
          type: string
      required:
      - publicKey
    LeaseUpdateRequest:
      type: object
      properties:
        agentId:
          type: string
        publicKey:
          type: string
        encryptedData:
          type: string
    UpdateSecretRequest:
      type: object
      properties:
        name:
          type: string
        encryptedValue:
          type: string
          maxLength: 87381
          minLength: 0
        metadata:
          type: object
          additionalProperties:
            type: object
        updatedLeases:
          type: array
          items:
            $ref: "#/components/schemas/LeaseUpdateRequest"
    UpdateRequestDTO:
      type: object
      properties:
        status:
          type: string
          enum:
          - pending
          - fulfilled
          - rejected
          - abandoned
        secretId:
          type: string
          pattern: "^[0-9]+$"
        rejectionReason:
          type: string
      required:
      - status
    LeaseResponse:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        agentName:
          type: string
        publicKey:
          type: string
        encryptedData:
          type: string
        expiry:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SecretResponse:
      type: object
      properties:
        secretId:
          type: string
        name:
          type: string
        encryptedValue:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentResponse:
      type: object
      properties:
        agentId:
          type: string
        name:
          type: string
        displayName:
          type: string
        publicKey:
          type: string
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT